Перемещает подписки на события в компоненты
