+++ src/utils/event.js
@@ -2,13 +2,6 @@ import dayjs from 'dayjs';
 import duration from 'dayjs/plugin/duration';
 dayjs.extend(duration);
 
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
 const humanizeEventDate = (dueDate) => dayjs(dueDate).format('D MMMM');
 const humanizeEventtime = (dueDate) => dayjs(dueDate).format('HH:mm');
 
@@ -22,6 +15,4 @@ const getTimeDuration = (dateTo, dateFrom) =>
     .replace(/^00H /, '')
     .replace(/^0/, '');
 
-const isEscEvent = (evt) => evt.key === 'Escape' || evt.key === 'Esc';
-
-export { getRandomInteger, humanizeEventDate, humanizeEventtime, humanizeDateAndTime, getTimeDuration, isEscEvent };
+export { humanizeEventDate, humanizeEventtime, humanizeDateAndTime, getTimeDuration};
